manipulator: restrict resizes to requested resizes
Example why this matters: if a disk partition is not trying to be
resized, and it's actual size is not an even multiple of what align_up
supports, then it could get a unrequested non-zero size change.
And if there is no supporting gap for that change, then we can get a
size Exception despite not having asked for a resize.